% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulations.R
\name{documents}
\alias{documents}
\title{Retrieve regulations.gov documents}
\usage{
documents(apikey = NULL, countsOnly = NULL, encoded = NULL,
  keywords = NULL, docType = NULL, docketID = NULL, docketType = NULL,
  commentPeriod = NULL, agency = NULL, nresults = 1000, offset = NULL,
  closingSoon = NULL, newlyPosted = NULL, comStartDate = NULL,
  comEndDate = NULL, createDate1 = NULL, createDate2 = NULL,
  receivedDate1 = NULL, receivedDate2 = NULL, postedDate1 = NULL,
  postedDate2 = NULL, category = NULL, sortBy = NULL, sortOrder = NULL,
  docketSubtype = NULL, docketSubSubtype = NULL, docSubtype = NULL, ...)
}
\arguments{
\item{apikey}{Access key for accessing the regulations.gov API.
Get one at https://regulationsgov.github.io/developers/key/.}

\item{countsOnly}{1 (will return only the document count for a search query) 0 (will return documents as well)}

\item{encoded}{1 (will accept Regulations.gov style encoded parameters) 0 (will not accept such encoded parameters)}

\item{keywords}{Keywords associated with the document}

\item{docType}{Document Type: N: Notice; PR: Proposed Rule; FR: Rule;
O: Other; SR: Supporting & Related Material; PS: Public Submission}

\item{docketID}{Valid Docket ID (ex. SEC-2012-0044)}

\item{docketType}{"Docket Type: R: Rulemaking; N: Nonrulemaking}

\item{commentPeriod}{Comment Period: O: Open; C: Closed}

\item{agency}{Federal Agency: List of accepted Federal Agency values. This field allows multiple values. ex: c(FMCSA, EPA, FDA)}

\item{nresults}{Results Per Page 10, 25, 100, 500, 1000.  Results per page may not exceed 1000.}

\item{offset}{Page offset (always starts with 0). This is used in
conjunction with results per page to provide large data sets. For example,
if a search produces 82 results and the result per page is set to 25, this
will generate 4 pages. 3 pages will have 25 results and the last page will
have 7 results. Page offset values for each page will be: Page 1: po=0
Page 2: po=25 Page 3: po=50 Page 4: po=75. The total number of pages
is [total results/results per page] and page offset for page X is
[X-1 * results per page]"}

\item{closingSoon}{Comment Period Closing Soon: 0 (closing today);
3 (closing within 3 days); 15 (closing within 15 days);
30 (closing within 30 days); 90 (closing within 90 days)"}

\item{newlyPosted}{Newly Posted: 0 (posted today);
3 (posted within last 3 days); 15 (posted within last 15 days);
30 (posted within last 30 days); 90 (posted within last 90 days).
For periods of time beyond 90-days, please use a date range with the Posted
Date parameter.}

\item{comStartDate}{Comment Period Start Date}

\item{comEndDate}{Comment Period End Date. Comment Period Start and
End Dates are mutually exclusive with the 'closing soon' parameter.
If both are provided, 'closing soon' will be ignored.}

\item{createDate1}{Beginning of date-of-creation range}

\item{createDate2}{End of date-of-creation range}

\item{receivedDate1}{Beginning of date-of-reception range}

\item{receivedDate2}{End of date-of-reception range}

\item{postedDate1}{Beginning of date-of-posting range}

\item{postedDate2}{End of date-of-posting range}

\item{category}{Document Category: AD (Aerospace and Transportation);
AEP (Agriculture, Environment, and Public Lands);
BFS (Banking and Financial); CT (Commerce and International);
LES (Defense, Law Enforcement, and Security);
EELS (Education, Labor, Presidential, and Government Services);
EUMM (Energy, Natural Resources, and Utilities);
HCFP (Food Safety, Health, and Pharmaceutical);
PRE (Housing, Development, and Real Estate);
ITT (Technology and Telecommunications)}

\item{sortBy}{Sort By: docketId (Docket ID); docId (Document ID);
title (Title); postedDate (Posted Date); agency (Agency);
documentType (Document Type); submitterName (Submitter Name);
organization (Organization). Sort Order is REQUIRED if this parameter is
included.}

\item{sortOrder}{Sort Order: ASC: Ascending; DESC: Descending}

\item{docketSubtype}{Docket Subtype: Only one docket subtype at a time may
be selected. One or more agency values must be part of the request.
Only values valid for the selected agency will be returned.}

\item{docketSubSubtype}{Docket Sub-subtype: Only one docket sub-subtype at
a time may be selected. One or more agency values must be part of the
request. Only values valid for the selected agency will be returned.}

\item{Document}{Subtype: Single or multiple document subtypes may be
included.  Multiple values should be passed as follows:
Certificate+of+Service%2BCorrespondence}
}
\description{
Retrieve regulations.gov documents
}
\details{
A Docket Type is either Rulemaking or Nonrulemaking. A Rulemaking docket includes the type of regulation that establishes a rule.
While a Non-Rulemaking docket does not include a rule.
}

